correct the path to sql db
Actually that has issues when using TS for instance, that database file will be 4 levels deep and you have to do ../../../../.tmp/data.db to get to it from the file's perspective if you are configuring this from an env like test. This happens because the actual file that will be used will result as `dist/config/env/test/database.js` for instance. so the db file is being created in `dist/config/.tmp/data.db`. There are just too many variables to use relative path. What needs to happen is just define the file path from the root of the project, hence the change proposed